+++ editor/changelog.md
@@ -3,6 +3,7 @@
 ## Solarus Quest Editor 2.0.1 (in progress)
 
 * Fix empty tileset selector (#1724).
+* Adjust selector size when expanding folders.
 
 ## Solarus Quest Editor 2.0.0 (2025-04-01)